Motor: Simplify macros
MOTORPATH_MAX was/is a macro that describes the maximum length of the
path to the motor directory in sysfs while MOTORPATHNAME_MAX is the
maximum length of the pathname of entries in that directory. Here, I
decide to just got with the bigger value also when the smaller could
be used and remove MOTORPATH_MAX because I find it uncessarily
confusing to have two macros.

Signed-off-by: Viktor Rosendahl <viktor.rosendahl@gmail.com>