Motor: Add more yet another batch of set/get
The readState() is supposed to be called before trying to determine
the state. The diea is that a call to readState() makes the Motor
object refresh it's internal state mask that can then be inspected
by isStateActive(), or alternatively the whole mask can be read with
getStateMask(). The bits in the masks can be obtained with
getStateMaskBit.

Signed-off-by: Viktor Rosendahl <viktor.rosendahl@gmail.com>